Use `IntoValue` for history config
