+++ crates/nu-protocol/src/config/history.rs
@@ -1,0 +1,20 @@
+use super::{prelude::*, HistoryFileFormat};
+
+#[derive(Clone, Copy, Debug, IntoValue, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HistoryConfig {
+    pub max_size: i64,
+    pub sync_on_enter: bool,
+    pub file_format: HistoryFileFormat,
+    pub isolation: bool,
+}
+
+impl Default for HistoryConfig {
+    fn default() -> Self {
+        Self {
+            max_size: 100_000,
+            sync_on_enter: true,
+            file_format: HistoryFileFormat::Plaintext,
+            isolation: false,
+        }
+    }
+}
